% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_disclosure_control.R
\name{check_column_types}
\alias{check_column_types}
\title{Check Column Types}
\usage{
check_column_types(file, pattern, config)
}
\arguments{
\item{file}{A .csv file to extract column types from.}

\item{pattern}{A character string of the file name pattern to check. This takes the form of a file name with extension and any dates removed.}

\item{config}{A list config file to check column types against. This file contains information about expected file metadata.}
}
\description{
Takes a .csv file and name pattern to determine if column types match as expected.
}
