% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_disclosure_control.R
\name{check_headers}
\alias{check_headers}
\title{Check Headers}
\usage{
check_headers(file, header, pattern, config)
}
\arguments{
\item{file}{An .html file to extract headers from.}

\item{header}{A character string of header type - either h1, h2, or h3.}

\item{pattern}{A character string of the file name pattern to check. This takes the form of a file name with extension and any dates removed.}

\item{config}{A list config file to check headers against. This file contains information about expected file metadata.}
}
\description{
Checking h1, h2 and h3 header elements match as expected.
}
