% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_disclosure_control.R
\name{check_number_of_rows}
\alias{check_number_of_rows}
\title{Check Number of Rows}
\usage{
check_number_of_rows(file, pattern, config)
}
\arguments{
\item{file}{A .csv file to extract number of rows from.}

\item{pattern}{A character string of the file name pattern to check. This takes the form of a file name with extension and any dates removed.}

\item{config}{A list config file to check number of rows against. This file contains information about expected file metadata.}
}
\description{
Takes a .csv file, name pattern and config file to determine if the number of rows match as expected.
}
