% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_disclosure_control.R
\name{check_number_of_tables}
\alias{check_number_of_tables}
\title{Check Number of Tables}
\usage{
check_number_of_tables(file, pattern, config)
}
\arguments{
\item{file}{An .html file to extract number of tables from.}

\item{pattern}{A character string of the file name pattern to check. This takes the form of a file name with extension and any dates removed.}

\item{config}{A list config file to check number of tables against. This file contains information about expected file metadata.}
}
\description{
Checking number of tables in a .html document match as expected.
}
