% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_disclosure_control.R
\name{check_yaml_file}
\alias{check_yaml_file}
\title{Check YAML File}
\usage{
check_yaml_file(file, pattern, config)
}
\arguments{
\item{file}{A .yaml file to check.}

\item{pattern}{A character string of the file name pattern to check. This takes the form of a file name with extension and any dates removed.}

\item{config}{A list config file to check file attributes against. This file contains information about expected file metadata.}
}
\description{
Function to run .yaml specific checks on a .yaml file
}
