% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_disclosure_control.R
\name{check_yaml_names}
\alias{check_yaml_names}
\title{Check YAML Names}
\usage{
check_yaml_names(file, pattern, config)
}
\arguments{
\item{file}{A .yaml file to extract names from.}

\item{pattern}{A character string of the file name pattern to check. This takes the form of a file name with extension and any dates removed.}

\item{config}{A list config file to check file attributes against. This file contains information about expected file metadata.}
}
\description{
Function to check that names in .yaml file match as expected.
}
