% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{clean_file}
\alias{clean_file}
\title{clean_file}
\usage{
clean_file(path)
}
\arguments{
\item{path}{path to clean}
}
\value{
character
}
\description{
function to remove extension from file path and remove all of path up to last path separator
}
