% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit-hooks-utils.R
\name{commit_hooks_setup}
\alias{commit_hooks_setup}
\title{Commit hooks setup}
\usage{
commit_hooks_setup(path_to_git_repo)
}
\arguments{
\item{path_to_git_repo}{}
}
\description{
this function establishes commit hooks to ensure that ipynb
outputs cannot be pushed to github
}
