% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{convert_log_to_df}
\alias{convert_log_to_df}
\title{convert_log_to_df}
\usage{
convert_log_to_df(file_path)
}
\arguments{
\item{file_path}{string path of the raw log file}
}
\description{
function to convert a log file to a data.frame
}
\examples{
# Load the library to ensure example data sets are available
library(gcptools)

# activate authentication token
# authenticate_GCP()

# Create dummy log file
# gcptools::initialise_log(log_file = "example",
#                         log_dir = "convert_log_to_df_example")

# file_path <- list.files("convert_log_to_df_example", full.names = TRUE)

# call function
# dummy_output <- convert_log_to_df(file_path)

# view output
# dummy_output

# remove folder/files used in the example
# unlink("convert_log_to_df_example/", recursive = TRUE)
}
