% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{count_warnings_and_errors}
\alias{count_warnings_and_errors}
\title{count_warnings_and_errors}
\usage{
count_warnings_and_errors(log_file_name)
}
\arguments{
\item{log_file_name}{the name of the log file}
}
\value{
invisible
}
\description{
Counts number of warnings and errors in a log file and appends it to the bottom of the log file
}
