% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_config}
\alias{create_config}
\title{create_config}
\usage{
create_config(...)
}
\arguments{
\item{...}{Either folder to read in, e.g. create_config("configuration", subfolder = "model")
or filepaths to read in, e.g. create_config("config.yaml", "config2.yaml")}
}
\value{
list
}
\description{
A function that reads in configuration settings from either a folder or file(s).
}
\seealso{
Other Config Creation Functions: 
\code{\link{create_config_from_files}()},
\code{\link{create_config_from_folder}()}
}
\concept{Config Creation Functions}
