% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_config_from_files}
\alias{create_config_from_files}
\title{create_config_from_files}
\usage{
create_config_from_files(...)
}
\arguments{
\item{...}{filepaths to read in, e.g. create_config("config.yaml", "config2.yaml")}
}
\value{
list
}
\description{
function to append config files to a single list
}
\seealso{
Other Config Creation Functions: 
\code{\link{create_config_from_folder}()},
\code{\link{create_config}()}
}
\concept{Config Creation Functions}
