% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{create_config_from_folder}
\alias{create_config_from_folder}
\title{create_config_from_folder}
\usage{
create_config_from_folder(folder, subfolder = "models")
}
\arguments{
\item{folder}{This is the path to the folder containing configuration settings}

\item{subfolder}{This is a subfolder within folder.
The contents of this subfolder will be recursively read in and flattened
and added to the output list as element "subfolder".}
}
\value{
list
}
\description{
A function to append config files to a single list from a folder.
}
\seealso{
Other Config Creation Functions: 
\code{\link{create_config_from_files}()},
\code{\link{create_config}()}
}
\concept{Config Creation Functions}
