% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automated_disclosure_control.R
\name{disco_checks}
\alias{disco_checks}
\title{Disco Checks}
\usage{
disco_checks(
  path,
  file,
  config,
  type = c(".csv", ".jpeg|.jpg", ".yaml|.yml", ".html")
)
}
\arguments{
\item{path}{A character string file path of file to be checked.}

\item{file}{File to be checked. Can take the form of a .csv, .html, .yaml or .jpeg.}

\item{config}{A list config file to check file attributes against. This file contains information about expected file metadata.}

\item{type}{A character string of the file type.}
}
\value{
Pass/fail messages
}
\description{
Takes a file and file path and determines whether or not the file passes automated disclosure checks.
}
