% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette.R
\name{display_vignette}
\alias{display_vignette}
\title{display_vignette}
\usage{
display_vignette(vignette, package)
}
\arguments{
\item{vignette}{The name of the vignette you wish to display, do not include .html at the end (string).}

\item{package}{The package containing the vignette you wish to display (string).}
}
\value{
invisible
}
\description{
This function will display a chosen vignette of any local ("cis") package in your notebook/console. N.b. It will not work in terminal.
}
