% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_report.R
\name{download_qa_report_to_notebook}
\alias{download_qa_report_to_notebook}
\title{download_qa_report_to_notebook}
\usage{
download_qa_report_to_notebook(file, bucket = "polestar-prod-process-wip")
}
\arguments{
\item{file}{full path to file}

\item{bucket}{the bucket containing the qa report (defaults to WIP)}
}
\description{
function to read in qa reports from bucket
}
