% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_csv}
\alias{dummy_csv}
\title{Dummy CSV}
\format{
An object of class \code{data.frame} with 1 rows and 8 columns.
}
\source{
created by data-raw/make_dummy_csv.R
}
\usage{
dummy_csv
}
\description{
A dummy csv file to run unit tests on csv disclosure control checking functions.
}
\keyword{datasets}
