% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_yaml}
\alias{dummy_yaml}
\title{Dummy YAML}
\format{
An object of class \code{list} of length 7.
}
\source{
created by data-raw/make_dummy_yaml.R
}
\usage{
dummy_yaml
}
\description{
A dummy yaml file to run unit tests on yaml disclosure control checking functions.
}
\keyword{datasets}
