% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycling_bin.R
\name{empty_recycling_bin}
\alias{empty_recycling_bin}
\title{Delete files from "Trash" folder}
\usage{
empty_recycling_bin()
}
\value{
invisible
}
\description{
This function is a wrapper around a system call to rm which empties the trash folder.
}
\examples{

# Call the function
empty_recycling_bin()

}
