% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_regions}
\alias{england_regions}
\title{England Regions}
\format{
An object of class \code{character} of length 9.
}
\source{
created by data-raw/make_geographies.R
}
\usage{
england_regions
}
\description{
A vector of the NUTS1 Region names in England
}
\keyword{datasets}
