% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure_control.R
\name{explore_column}
\alias{explore_column}
\title{explore_column}
\usage{
explore_column(data, column_name)
}
\arguments{
\item{data}{a dataframe containing the data you want to explore}

\item{column_name}{a string containing the exact name (case-sensitive) of the column you want to explore}
}
\value{
either unique values or a percentage of low counts under a given threshold (default 10)
}
\description{
function for use in disclosure control to explore the contents of a particular column
}
