% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{filter_log_by_date}
\alias{filter_log_by_date}
\title{filter_log_by_date}
\usage{
filter_log_by_date(file_list, data_run_to, data_run_from)
}
\arguments{
\item{file_list}{data.frame all log files located in the bucket}

\item{data_run_to}{This is the latest date (inclusive) you want logs for (set as NULL if you don't want to restrict logs to before a selected date.}

\item{data_run_from}{This is the earliest date (inclusive) you want logs for (set as NULL if you don't want to restrict logs to after a selected date.}
}
\value{
files_to_download list of files that meet date criteria
}
\description{
filter log files by specific dates
}
