% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{filter_log_by_pipeline}
\alias{filter_log_by_pipeline}
\title{filter_log_by_pipeline}
\usage{
filter_log_by_pipeline(
  file_list,
  pipeline = c("mrp", "incidence", "inla", "age_contour", "all")
)
}
\arguments{
\item{file_list}{data.frame all log files located in the bucket}

\item{pipeline}{string set specific pipeline or all pipelines}
}
\value{
files_to_download list of files that meet date criteria
}
\description{
filter log files by specific pipeline i.e. "mrp", "incidence", "inla", "age_contour", "all"
}
