% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{filter_log_level}
\alias{filter_log_level}
\title{filter_log_level}
\usage{
filter_log_level(log, level)
}
\arguments{
\item{log}{data frame}

\item{level}{Levels to be filtered should be a character vector of any number of the following strings "INFO", "TRACE", "DEBUG", "WARN", "ERROR", "FATAL"

return filtered data frame}
}
\description{
function to return specific levels
}
