% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/gcp_utils.R
\docType{data}
\name{gcp_paths}
\alias{gcp_paths}
\title{Google Cloud Platform Bucket Paths}
\format{
An object of class \code{list} of length 3.

An object of class \code{list} of length 3.
}
\source{
created by data-raw/make_gcp_paths.R

NA
}
\usage{
gcp_paths

gcp_paths
}
\description{
Bucket paths to the Polestar Project work areas within google cloud platform

A list containing paths for three buckets on polestar, data, wip and review
   wip_bucket = "ons-psplus-data-prod-psplus-cis-data",
   data_bucket = "ons-psplus-data-prod-psplus-cis-data",
   review_bucket = "ons-psplus-analysis-prod-cis-review"
}
\keyword{datasets}
