% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp_read_object.R
\name{gcp_read_as_raw}
\alias{gcp_read_as_raw}
\title{gcp_read_as_raw}
\usage{
gcp_read_as_raw(object)
}
\arguments{
\item{object}{the object to read in - not needed to specify this when using as a parse function in #'gcs_get_objects}
}
\value{
invisible as this is only used as a parse function in gcs_get_objects
}
\description{
a parse function supplied to gcs_get_object to read in a data object in a .raw format.
}
