% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp_read_object.R
\name{gcp_read_csv}
\alias{gcp_read_csv}
\title{gcp_read_csv}
\usage{
gcp_read_csv(
  file_paths,
  bucket = c("wip_bucket", "data_bucket", "review_bucket"),
  col_types = NULL
)
}
\arguments{
\item{file_paths}{one or more file_paths to be read in. If more than one file path is provided, it is easier to provide this as a named list as the data will be returned with the same names.}

\item{bucket}{GCP bucket to be read in from (default is data bucket)}

\item{col_types}{a list of column names and col_types across all files you are reading in}
}
\value{
a dataframe if one file path is provided or a list of dataframes if multiple file paths are provided
}
\description{
function to read in csv files from the GCP buckets. This function is a wrapper for the gcp_read_single_csv and gcp_read_multiple_csvs functions.
}
