% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp_read_object.R
\name{gcp_read_rda}
\alias{gcp_read_rda}
\title{gcp_load_rda}
\usage{
gcp_read_rda(file, bucket = gcptools::gcp_paths$data_bucket)
}
\arguments{
\item{file}{rda object to read in}
}
\description{
function to read in rda files, N.B this currently only works if the .Rda file contains only a single object
}
