% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp_read_object.R
\name{gcp_read_rdata}
\alias{gcp_read_rdata}
\title{gcp_read_rdata}
\usage{
gcp_read_rdata(
  filepath = "inla_graphs",
  bucket = gcptools::gcp_paths$data_bucket
)
}
\arguments{
\item{filepath}{a string containing file path (including folder name) to the yaml file which you want to read in}

\item{bucket}{a string containing the name of the gcp bucket you want to read the yaml file from}
}
\value{
a data object which you have read in from the bucket
}
\description{
function to read rdata files in to the environment from a gcpbucket
}
