% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp_read_object.R
\name{gcp_read_single_csv}
\alias{gcp_read_single_csv}
\title{gcp_read_single_csv}
\usage{
gcp_read_single_csv(file_path, bucket, col_types = NULL)
}
\arguments{
\item{file_path}{- the file path to be read in.}

\item{bucket}{GCP bucket to be read in from (default is data bucket)}

\item{col_types}{a list of column names and col_types across the file you are reading in.}
}
\value{
a dataframe
}
\description{
function to read in a single csv file from a GCP bucket. This function is used inside the #'gcp_read_csv wrapper.
}
