% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp_read_object.R
\name{gcp_read_yaml}
\alias{gcp_read_yaml}
\title{gcp_read_yaml}
\usage{
gcp_read_yaml(filepath, bucket = gcp_paths$review_bucket)
}
\arguments{
\item{filepath}{a string containing file path (including folder name) to the yaml file which you want to read in}

\item{bucket}{a string containing the name of the gcp bucket you want to read the yaml file from}
}
\value{
a list containing the contents of the yaml file
}
\description{
function to read yaml files in to the environment from a gcpbucket
}
