% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp_upload_object.R
\name{gcp_upload_csv}
\alias{gcp_upload_csv}
\title{gcp_upload_csv}
\usage{
gcp_upload_csv(data, filename, bucket)
}
\arguments{
\item{data}{data to upload}

\item{filename}{desired output filename including folder}

\item{bucket}{desired output bucket}
}
\value{
None (file uploaded to bucket)
}
\description{
gcp_upload_csv
}
