% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp_upload_object.R
\name{gcp_upload_object}
\alias{gcp_upload_object}
\title{gcp_upload_object}
\usage{
gcp_upload_object(object, filepath, file_type, bucket)
}
\arguments{
\item{object}{a named object within the environment which you want to upload}

\item{filepath}{the desired filename and folder name combined as one string}

\item{file_type}{a string containing the file extension without the '.' at the start e.g. 'csv'}

\item{bucket}{desired output bucket}
}
\value{
None (file uploaded to bucket)
}
\description{
gcp_upload_object
}
