% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp_upload_object.R
\name{gcp_upload_raw_log}
\alias{gcp_upload_raw_log}
\title{gcp_upload_raw_log}
\usage{
gcp_upload_raw_log(
  file_path,
  destination_folder = "logs/raw/",
  pipeline = c("mrp", "incidence", "inla", "age_contour"),
  bucket = "polestar-prod-process-wip",
  config
)
}
\arguments{
\item{file_path}{path to the raw data}

\item{destination_folder}{folder name in bucket}

\item{pipeline}{desired output pipeline name}

\item{bucket}{desired output bucket}

\item{config}{configuration settings read in by config.yaml}
}
\value{
None (file uploaded to bucket)
}
\description{
gcp_upload_raw_log
}
