% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp_write_object.R
\name{gcp_write_log}
\alias{gcp_write_log}
\title{gcp_write_log}
\usage{
gcp_write_log(input, output)
}
\arguments{
\item{input}{object to be read in log format}

\item{output}{log file}
}
\description{
function to read in log files
}
