% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_coverage_report.R
\name{generate_coverage_report}
\alias{generate_coverage_report}
\title{Generate coverage report}
\usage{
generate_coverage_report(dir)
}
\arguments{
\item{dir}{A string to be included in the output file name along with the date (usually the package name)}
}
\value{
a list of two elements filecoverage (each module) and totalcoverage (whole_package)
}
\description{
This function takes a file path to a package directory (e.g. cismrp). It creates a folder called coverage_report_<dir>_<yyyymmdd>.
             It runs covr::package_coverage and then generates a file report for each module in the package. It returns a list containing the
             test coverage for each module and for the package overall that can be added to an integrated test report
}
