% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_column_types}
\alias{get_column_types}
\title{get column types}
\usage{
get_column_types(config)
}
\arguments{
\item{config}{contains directory information on this occasion}
}
\value{
evaluated string of column names and types
}
\description{
reading column types from config and returning an evaluated string
}
