% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_get_functions.R
\name{get_current_branch}
\alias{get_current_branch}
\title{get_current_branch}
\usage{
get_current_branch()
}
\value{
a list containing the name, type and repo of the currently checked out branch
}
\description{
retrieve the current branch
}
