% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure_control.R
\name{get_filepaths_from_project_name}
\alias{get_filepaths_from_project_name}
\title{get_filepaths_from_project_name}
\usage{
get_filepaths_from_project_name(
  project_name,
  file_type = c(".csv", ".jpeg|.jpg", ".yaml|.yml", ".html")
)
}
\arguments{
\item{project_name}{a string containing the exact project name (case-sensitive) which you want to get the filepaths from}
}
\value{
dataframe containing the filepaths of all of the files within that project
}
\description{
function to search all of the available filepaths witihin a given project_name and return it as a dataframe
}
