% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_get_functions.R
\name{get_list_of_commits}
\alias{get_list_of_commits}
\title{get_list_of_commits}
\usage{
get_list_of_commits()
}
\value{
list of commits
}
\description{
retrieve a list of commits for a specific repository
}
