% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_get_functions.R
\name{get_log_dataframe_list}
\alias{get_log_dataframe_list}
\title{get_log_dataframe_list}
\usage{
get_log_dataframe_list(files_to_download, top_level_path, bucket)
}
\arguments{
\item{files_to_download}{list names of the logs files we want to download}

\item{top_level_path}{string path to search for log files}

\item{bucket}{desired output bucket}
}
\value{
log_list list of dataframe objects
}
\description{
reading in list of log file names and downloads them into a list of dataframe objects
}
\examples{
# Load the library to ensure the example data sets are available
# library(gcptools)

# Set up example variables to test the code
# files_to_download <- data.frame(filename = c("inla_20211213_processed_log.csv",
#                                              "mrp_20211213_processed_log.csv"))
top_level_path <- "logs/processed/"
# bucket <- "polestar-prod-process-wip"

# After loading in the variables you can call the function
# log_list <- get_log_dataframe_list(files_to_download, top_level_path, bucket)

# View the output
# log_list

}
