% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp_upload_object.R
\name{get_object_write_function}
\alias{get_object_write_function}
\title{get_object_write_function}
\usage{
get_object_write_function(
  file_type = c("rdata", "yaml", "jpeg", "html", "csv", "xlsx", "rds")
)
}
\arguments{
\item{file_type}{a string containing the file type extension e.g. 'csv'}
}
\value{
a write function
}
\description{
function to select an appropriate write function to parse by filetype
}
