% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_get_functions.R
\name{get_package_version}
\alias{get_package_version}
\title{get_package_version}
\usage{
get_package_version(package)
}
\arguments{
\item{package}{string input for the package you intend to check the version of}
}
\value{
outputs version number of package as string
}
\description{
checks the version of the package you are using
}
