% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure_control.R
\name{get_unique_values}
\alias{get_unique_values}
\title{get_unique_values}
\usage{
get_unique_values(data, column_name)
}
\arguments{
\item{data}{a dataframe containing the data you want to explore}

\item{column_name}{a string containing the exact name (case-sensitive) of the column you want to explore}
}
\value{
a pretty list containing the unique values within that column
}
\description{
function for use in disclosure control to explore the contents of a character or factor type column
}
