% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcp_utils.R
\name{gsutil_wrapper}
\alias{gsutil_wrapper}
\title{Move files from notebook to bucket}
\usage{
gsutil_wrapper(
  input,
  output,
  bucket = c("wip_bucket", "review_bucket"),
  keep.local = FALSE
)
}
\arguments{
\item{input}{The name of the object to upload to the bucket.}

\item{output}{The desired name to save the object to in the bucket (accepts '/'). It currently only works with ".csv"s.}

\item{bucket}{The bucket to save to one of 'wip_bucket' or 'review_bucket'}

\item{keep.local}{Boolean (defaults FALSE) controls whether th interim file where the object is saved to the notebook disk is retained.}
}
\value{
invisible
}
\description{
This function is a wrapper around a system call to gsutil mv/cp which moves / copies a file from your notebook to either the WIP or the review bucket.
}
\examples{

# create a test data frame
test_data <- data.frame(sample_column <- 1:10)

# Call the function with correct parameters
gsutil_wrapper(
  input = test_data,
  output = "test_data.csv",
  bucket = "wip_bucket",
  keep.local = FALSE
)

}
