% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_utils.R
\name{initialise_log}
\alias{initialise_log}
\title{initialise_log}
\usage{
initialise_log(
  log_dir,
  log_file,
  log_threshold = c("TRACE", "DEBUG", "INFO", "WARN", "ERROR", "FATAL")
)
}
\arguments{
\item{log_dir}{the directory to store log files (if missing defaults to current working directory)}

\item{log_file}{the output file name}

\item{log_threshold}{one of "TRACE", "DEBUG", "INFO", "WARN", "ERROR", "FATAL"}
}
\value{
log_file_name the file path for the log file that is created
}
\description{
Opens up a new log file
}
