% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycling_bin.R
\name{list_recycling_bin}
\alias{list_recycling_bin}
\title{list_recycling_bin}
\usage{
list_recycling_bin()
}
\value{
data.frame
}
\description{
This function lists the contents of the recycling bin.
}
\examples{

# Call the function
list_recycling_bin()

}
