% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_model_info.R
\name{log_age_contour_model_info}
\alias{log_age_contour_model_info}
\title{log_age_contour_model_info}
\usage{
log_age_contour_model_info(config)
}
\arguments{
\item{config}{the config variable (after it has been loaded in as a list using load_config)}
}
\value{
invisible
}
\description{
Logs the username, analyst name and model information for the age contour pipeline
}
