% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_model_outputs.R
\name{log_age_contour_outputs}
\alias{log_age_contour_outputs}
\title{log_age_contour_outputs}
\usage{
log_age_contour_outputs(config, time_stamp)
}
\arguments{
\item{config}{config list read in by load_config}

\item{time_stamp}{time_stamp generated in main}
}
\description{
log whether expected outputs are saved to WIP and review buckets for the age contour pipeline
}
