% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_model_info.R
\name{log_incidence_model_info}
\alias{log_incidence_model_info}
\title{log_incidence_model_info}
\usage{
log_incidence_model_info(main_config, country_configs)
}
\arguments{
\item{main_config}{the main config variable (after it has been loaded in as a list using load_config)}

\item{country_configs}{the country_configs list (after they have been loaded in using load_country_configs)}
}
\value{
invisible
}
\description{
Logs the analyst name and model information for the incidence pipeline
}
