% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_model_outputs.R
\name{log_incidence_outputs}
\alias{log_incidence_outputs}
\title{log_incidence_outputs}
\usage{
log_incidence_outputs(config)
}
\arguments{
\item{config}{main config read in by load_config}
}
\description{
logs whether expected outputs are saved to WIP and review buckets for the Incidence pipeline
}
