% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_model_info.R
\name{log_inla_model_info}
\alias{log_inla_model_info}
\title{log_inla_model_info}
\usage{
log_inla_model_info(config)
}
\arguments{
\item{config}{the config variable (after it has been loaded in as a list using load_config)}
}
\value{
invisible
}
\description{
Logs the analyst name and model information for the INLA pipeline
}
