% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_model_outputs.R
\name{log_inla_outputs}
\alias{log_inla_outputs}
\title{log_inla_outputs}
\usage{
log_inla_outputs(config)
}
\arguments{
\item{config}{config as read in by load_config}
}
\description{
logs whether expected outputs are saved to WIP and review buckets for the INLA pipeline
}
