% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_model_info.R
\name{log_model_info}
\alias{log_model_info}
\title{log_model_info}
\usage{
log_model_info(
  main_config,
  pipeline = c("mrp", "age_contour", "incidence", "inla"),
  country_configs = NULL
)
}
\arguments{
\item{main_config}{the main config variable (after it has been loaded in as a list using load_config)}

\item{pipeline}{the pipeline that you are running - this should be either mrp, age_contour, incidence or inla}

\item{country_configs}{the country_configs list (after they have been loaded in using load_country_configs) - this is required for the mrp and incidence pipelines but optional for other pipelines}
}
\value{
invisible
}
\description{
this function logs the username of the user and based on pipeline specified will run either the log_mrp_model_info, log_age_contour_model_info, log_incidence_model_info or log_inla_model_info functions, which will log model information for their respective models
}
