% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_model_outputs.R
\name{log_mrp_outputs}
\alias{log_mrp_outputs}
\title{log_mrp_outputs}
\usage{
log_mrp_outputs(region_diffs, config, time_stamp)
}
\arguments{
\item{region_diffs}{region_diffs generated in main.R}

\item{config}{main config read in by load_config}

\item{time_stamp}{timestamp of when outputs were saved}
}
\description{
logs whether expected outputs are saved to WIP and review buckets for the MRP pipeline
}
