% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_model_outputs.R
\name{log_output_files}
\alias{log_output_files}
\title{log_output_files}
\usage{
log_output_files(
  config,
  expected_wip_files = NULL,
  expected_review_files = NULL,
  folder_path
)
}
\arguments{
\item{config}{config list read in by load_config}

\item{expected_wip_files}{a named vector or list containing files that are expected to be saved to the wip bucket - leave blank or set as NULL if you don't want to log outputs that are saved to the WIP bucket}

\item{expected_review_files}{a named vector or list containing files that are expected to be saved to the review bucket - leave blank or set as NULL if you don't want to log outputs that are saved to the review bucket}

\item{folder_path}{the folder path where the files are expected to be saved}
}
\description{
logs whether expected files have been saved to the wip and review buckets
}
