% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuts1_to_country}
\alias{nuts1_to_country}
\title{NUTS1 Regions to Country Lookup}
\format{
An object of class \code{character} of length 12.
}
\source{
created by data-raw/make_geographies.R
}
\usage{
nuts1_to_country
}
\description{
A named vector of NUTS1 regions with their associated country as a value
}
\keyword{datasets}
