% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons_logo.R
\name{ons_logo}
\alias{ons_logo}
\title{ONS Logo.}
\usage{
ons_logo()
}
\value{
url to logo.
}
\description{
Base 64 code for the ONS logo for insertion in the table of contents (TOC) for Rmarkdown documents HTML.
}
