% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure_control.R
\name{print_jpeg}
\alias{print_jpeg}
\title{print_jpeg}
\usage{
print_jpeg(jpeg_filepath, bucket = gcp_paths$review_bucket)
}
\arguments{
\item{jpeg_filepath}{a string containing a jpeg file path (including folder name) that you want to print to the console}

\item{bucket}{the bucket which you want to read the jpeg from}
}
\value{
an image of the jpeg within the console
}
\description{
function to print jpegs to the console
}
