% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycling_bin.R
\name{recover_recycling_bin}
\alias{recover_recycling_bin}
\title{Recover files from "Trash" folder}
\usage{
recover_recycling_bin(recovery_dir = "recovered")
}
\arguments{
\item{recovery_dir}{Name of new directory to hold contents of Trash/files (defaults to "recovered").}
}
\value{
invisible
}
\description{
This function is a wrapper around a system call to mv which transfers files from the trash folder to the user top level.
}
\examples{

# Call the function
recover_recycling_bin()

}
