% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure_control.R
\name{retrieve_html_file}
\alias{retrieve_html_file}
\title{retrieve_html_file}
\usage{
retrieve_html_file(html_filepath)
}
\arguments{
\item{html_filepath}{a string containing a html file path (including folder name) that you want to save to the QA_reports folder}
}
\value{
a message saying where to find the html file
}
\description{
function to retrieve html file from the review bucket and put it in the QA_reports folder
}
