% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure_control.R
\name{reveal_low_counts}
\alias{reveal_low_counts}
\title{reveal_low_counts}
\usage{
reveal_low_counts(data, column_name, low_count_threshold = 10)
}
\arguments{
\item{data}{a dataframe containing the data you want to explore}

\item{column_name}{a string containing the exact name (case-sensitive) of the column you want to explore}

\item{low_count_threshold}{the threshold at which counts are considered disclosive (default 10)}
}
\value{
a message to the console which states the percentage of low counts within the column.
}
\description{
function for use in disclosure control to explore the percentage of low counts within a given numeric column
}
