% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{save_html}
\alias{save_html}
\title{save_html}
\usage{
save_html(input, output)
}
\arguments{
\item{input}{object to be read in html format}

\item{output}{html file}
}
\description{
function to read in html files
}
