% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_report.R
\name{save_log_report}
\alias{save_log_report}
\title{save_log_report}
\usage{
save_log_report(
  report,
  destination_folder,
  pipeline,
  data_run_to,
  data_run_from,
  bucket
)
}
\arguments{
\item{report}{csv file containing stitch logs}

\item{destination_folder}{folder to write report}

\item{pipeline}{string set specific pipeline or all pipelines}

\item{data_run_to}{This is the latest date (inclusive) you want logs for (set as NULL if you don't want to restrict logs to before a selected date).}

\item{data_run_from}{This is the earliest date (inclusive) you want logs for (set as NULL if you don't want to restrict logs to after a selected date).}

\item{bucket}{desired output bucket}
}
\value{
None (file uploaded to bucket)
}
\description{
saves report to the bucket
}
