% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_report.R
\name{save_qa_report}
\alias{save_qa_report}
\title{save QA report}
\usage{
save_qa_report(
  html_file_to_read,
  bucket = "polestar-prod-process-wip",
  config,
  filename
)
}
\arguments{
\item{html_file_to_read}{path to qa report, includes object name}

\item{bucket}{either review_bucket or wip_bucket}

\item{config}{config list containing the output path}

\item{filename}{which the qa report will be saved in the bucket}
}
\description{
save QA report to the selected bucket
}
