% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_report.R
\name{save_qa_report_age_contour}
\alias{save_qa_report_age_contour}
\title{save QA report}
\usage{
save_qa_report_age_contour(html_file_to_read, config, time_stamp)
}
\arguments{
\item{html_file_to_read}{path to qa report, includes object name}

\item{config}{config list containing the output path}

\item{time_stamp}{the time_stamp of this model run}
}
\description{
save QA report to the review and wip buckets
}
