% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uk_countries}
\alias{uk_countries}
\title{United Kingdom Countries}
\format{
An object of class \code{character} of length 4.
}
\source{
created by data-raw/make_geographies.R
}
\usage{
uk_countries
}
\description{
The names of the countries of the United Kingdom
}
\keyword{datasets}
