% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_report.R
\name{write_log_report_to_bucket}
\alias{write_log_report_to_bucket}
\title{write_log_report_to_bucket}
\usage{
write_log_report_to_bucket(
  pipeline = c("mrp", "incidence", "inla", "age_contour", "all"),
  top_level_path = "logs/processed/",
  data_run_from = NULL,
  data_run_to = NULL,
  destination_folder = "logs/report/",
  bucket = "polestar-prod-process-wip"
)
}
\arguments{
\item{pipeline}{string set specific pipeline or all pipelines}

\item{top_level_path}{string path to search for log files}

\item{data_run_from}{string setting specific date to run from}

\item{data_run_to}{setting string specific date to run to}

\item{destination_folder}{folder to write report}

\item{bucket}{desired output bucket}
}
\value{
None (file uploaded to bucket)
}
\description{
reads in processed logs combines them into a single csv file and
saves that single file back to the bucket
}
